+++ npm/brandup-ui-richeditor/source/index.ts
@@ -1,11 +1,15 @@
 export { default } from "./richeditor";
 export * from "./richeditor";
 export {
+	ALL_EDITOR_ACTIONS,
 	ALL_FORMAT_TOOLS,
+	EDITOR_ACTIONS,
 	FORMAT_TOOLS,
+	parseEditorActions,
 	parseFormatTools,
 	defaultFormatMarkers,
 	normalizeWhitespace,
+	type EditorAction,
 	type FormatTool,
 	type FormatStorage,
 	type FormatMarkers,